cn10k: rpm_intf: support for negative tx in tune serdes
Updated ETH_CMD_TUNE_SERDES command by adding support
for negative values of pre2, pre1 and post Tx Eq.
Negative values can be passed by using two's
complement of an 8-bit absolute value of pre2, pre1 or
post.

Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/boot/atf/+/88949
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>

+++ drivers/marvell/octeontx/cn10k/rpm_intf.c
@@ -1855,9 +1855,9 @@ static int rpm_process_requests(int rpm_id, int lmac_id)
 			ret = rpm_set_serdes_tx_tune(
 				 scratchx1.s.gser_tune.portm_idx,
 				 scratchx1.s.gser_tune.tx_main,
-				 scratchx1.s.gser_tune.tx_pre,
-				 scratchx1.s.gser_tune.tx_pre2,
-				 scratchx1.s.gser_tune.tx_post,
+				 (int8_t)scratchx1.s.gser_tune.tx_pre,
+				 (int8_t)scratchx1.s.gser_tune.tx_pre2,
+				 (int8_t)scratchx1.s.gser_tune.tx_post,
 						     lmac_ctx);
 			break;
